requirements.txt
boto3==1.9.191
botocore==1.12.191
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
public==2019.4.13
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
temp==2019.4.13
urllib3==1.25.3
Werkzeug==0.15.5
